Add ProcessExecutor service and impl, and move SelfDescriptionServiceImpl
