+++ src/main/java/com/tecnalia/datausage/service/ProcessExecutor.java
@@ -1,0 +1,8 @@
+package com.tecnalia.datausage.service;
+
+import java.util.List;
+
+public interface ProcessExecutor {
+
+	String executeProcess(List<String> cmdList);
+}
